+++ sider__exttest.py
@@ -1,0 +1,5 @@
+import sidertests
+
+ext_loaded = 'yes'
+sidertests.exttest_count += 1
+